revews page card modify
